telegram/requirements.txt
flask
playwright
requests
chromadb
python-dotenv
uvicorn
langchain-groq
dotenv
langchain
requests
telethon
sentence_transformers
flask_cors
praw
asyncio
aiohttp